Updates to source dropping implementation.
Added source propagation flag that only performs monitoring starting with set
tag events. This is done by inverting the sense of the flag array/cache. For
source propagation, 0 is invalid and 1 is valid. The flag is set to valid on a
not-dropped set tag event.

With source propagation, stores in UMC are labeled as set tag operations.

Moved set tag events into their own category for statistics.